% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra-package.R
\docType{package}
\name{xtra-package}
\alias{xtra}
\alias{xtra-package}
\title{xtra: Extra Utility Functions for R}
\description{
A growing collection of extra utility functions for R. Current tools include reading Excel files that use multiple header rows with merged cells, saving plots to disk, and adding images with optional captioning to Word documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NourEdinDarwish/xtra}
  \item \url{https://nouredindarwish.github.io/xtra/}
  \item Report bugs at \url{https://github.com/NourEdinDarwish/xtra/issues}
}

}
\author{
\strong{Maintainer}: Nour Edin Darwish \email{nouredindarwish@gmail.com} (\href{https://orcid.org/0009-0009-5527-4539}{ORCID})

}
\keyword{internal}
