% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtra-package.R
\docType{package}
\name{xtra-package}
\alias{xtra}
\alias{xtra-package}
\title{xtra: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NourEdinDarwish/xtra}
  \item Report bugs at \url{https://github.com/NourEdinDarwish/xtra/issues}
}

}
\author{
\strong{Maintainer}: Nour Edin Darwish \email{nouredindarwish@gmail.com} (\href{https://orcid.org/0009-0009-5527-4539}{ORCID})

}
\keyword{internal}
